pyproject.toml
[project]
name = "alloca"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "celery[redis]>=5.6.3",
    "fastapi>=0.136.1",
    "geoalchemy2>=0.19.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"